Revert add sdlc server to legend query application (#1405)
