+++ packages/legend-application-query-bootstrap/scripts/setup.js
@@ -49,7 +49,6 @@ export const setup = (outputDir) => {
         },
         studio: {
           url: 'http://localhost:8080/studio',
-          sdlcUrl: 'http://localhost:7070/api',
         },
       },
       undefined,
