+++ packages/legend-application-query/src/application/LegendQueryApplicationConfig.ts
@@ -66,7 +66,7 @@ export interface LegendQueryApplicationConfigurationData
     TEMPORARY__useLegacyDepotServerAPIRoutes?: boolean;
   };
   engine: { url: string; queryUrl?: string };
-  studio: { url: string; sdlcUrl: string };
+  studio: { url: string };
 }
 
 export class LegendQueryApplicationConfig extends LegendApplicationConfig {
@@ -76,7 +76,6 @@ export class LegendQueryApplicationConfig extends LegendApplicationConfig {
   readonly engineQueryServerUrl?: string | undefined;
   readonly depotServerUrl: string;
   readonly studioUrl: string;
-  readonly sdlcUrl: string;
   readonly TEMPORARY__useLegacyDepotServerAPIRoutes?: boolean | undefined;
 
   constructor(
@@ -100,10 +99,6 @@ export class LegendQueryApplicationConfig extends LegendApplicationConfig {
     this.studioUrl = guaranteeNonEmptyString(
       input.configData.studio.url,
       `Can't configure application: 'studio.url' field is missing or empty`,
-    );
-    this.sdlcUrl = guaranteeNonEmptyString(
-      input.configData.studio.sdlcUrl,
-      `Can't configure application: 'studio.sdlcUrl' field is missing or empty`,
     );
     this.TEMPORARY__useLegacyDepotServerAPIRoutes =
       input.configData.depot.TEMPORARY__useLegacyDepotServerAPIRoutes;
