+++ packages/legend-application-query/src/components/LegendQueryBaseStoreProvider.tsx
@@ -26,7 +26,6 @@ import {
   type ApplicationStore,
   useApplicationStore,
 } from '@finos/legend-application';
-import { useSDLCServerClient } from '@finos/legend-server-sdlc';
 
 export const useLegendQueryApplicationStore = (): ApplicationStore<
   LegendQueryApplicationConfig,
@@ -47,13 +46,11 @@ export const LegendQueryBaseStoreProvider: React.FC<{
 }> = ({ children, pluginManager }) => {
   const applicationStore = useLegendQueryApplicationStore();
   const depotServerClient = useDepotServerClient();
-  const sdlcServerClient = useSDLCServerClient();
   const store = useLocalObservable(
     () =>
       new LegendQueryBaseStore(
         applicationStore,
         depotServerClient,
-        sdlcServerClient,
         pluginManager,
       ),
   );
