+++ packages/legend-application-query/src/components/QuerySetupStoreProvider.tsx
@@ -23,7 +23,6 @@ import {
   useLegendQueryApplicationStore,
   useLegendQueryBaseStore,
 } from './LegendQueryBaseStoreProvider.js';
-import { useSDLCServerClient } from '@finos/legend-server-sdlc';
 
 const QuerySetupStoreContext = createContext<QuerySetupStore | undefined>(
   undefined,
@@ -35,13 +34,11 @@ export const QuerySetupStoreProvider: React.FC<{
   const applicationStore = useLegendQueryApplicationStore();
   const depotServerClient = useDepotServerClient();
   const baseStore = useLegendQueryBaseStore();
-  const sdlcServerClient = useSDLCServerClient();
   const store = useLocalObservable(
     () =>
       new QuerySetupStore(
         applicationStore,
         depotServerClient,
-        sdlcServerClient,
         baseStore.pluginManager,
       ),
   );
