+++ packages/legend-application-query/src/stores/QueryEditorStore.ts
@@ -68,7 +68,6 @@ import { TAB_SIZE, APPLICATION_EVENT } from '@finos/legend-application';
 import type { LegendQueryPluginManager } from '../application/LegendQueryPluginManager.js';
 import { LegendQueryEventService } from './LegendQueryEventService.js';
 import type { LegendQueryApplicationStore } from './LegendQueryBaseStore.js';
-import type { SDLCServerClient } from '@finos/legend-server-sdlc';
 
 export interface QueryExportConfiguration {
   defaultName?: string | undefined;
@@ -200,7 +199,6 @@ export class QueryExportState {
 export abstract class QueryEditorStore {
   applicationStore: LegendQueryApplicationStore;
   depotServerClient: DepotServerClient;
-  sdlcServerClient: SDLCServerClient;
   pluginManager: LegendQueryPluginManager;
   graphManagerState: GraphManagerState;
 
@@ -211,7 +209,6 @@ export abstract class QueryEditorStore {
   constructor(
     applicationStore: LegendQueryApplicationStore,
     depotServerClient: DepotServerClient,
-    sdlcServerClient: SDLCServerClient,
     pluginManager: LegendQueryPluginManager,
   ) {
     makeObservable(this, {
@@ -233,7 +230,6 @@ export abstract class QueryEditorStore {
       this.graphManagerState,
       new StandardQueryBuilderMode(),
     );
-    this.sdlcServerClient = sdlcServerClient;
   }
 
   setExportState(val: QueryExportState | undefined): void {
@@ -408,7 +404,6 @@ export class CreateQueryEditorStore extends QueryEditorStore {
   constructor(
     applicationStore: LegendQueryApplicationStore,
     depotServerClient: DepotServerClient,
-    sdlcServerClient: SDLCServerClient,
     pluginManager: LegendQueryPluginManager,
     groupId: string,
     artifactId: string,
@@ -417,7 +412,7 @@ export class CreateQueryEditorStore extends QueryEditorStore {
     runtimePath: string,
     classPath: string | undefined,
   ) {
-    super(applicationStore, depotServerClient, sdlcServerClient, pluginManager);
+    super(applicationStore, depotServerClient, pluginManager);
 
     this.groupId = groupId;
     this.artifactId = artifactId;
@@ -503,7 +498,6 @@ export class ServiceQueryEditorStore extends QueryEditorStore {
   constructor(
     applicationStore: LegendQueryApplicationStore,
     depotServerClient: DepotServerClient,
-    sdlcServerClient: SDLCServerClient,
     pluginManager: LegendQueryPluginManager,
     groupId: string,
     artifactId: string,
@@ -511,7 +505,7 @@ export class ServiceQueryEditorStore extends QueryEditorStore {
     servicePath: string,
     executionKey: string | undefined,
   ) {
-    super(applicationStore, depotServerClient, sdlcServerClient, pluginManager);
+    super(applicationStore, depotServerClient, pluginManager);
 
     this.groupId = groupId;
     this.artifactId = artifactId;
@@ -595,11 +589,10 @@ export class ExistingQueryEditorStore extends QueryEditorStore {
   constructor(
     applicationStore: LegendQueryApplicationStore,
     depotServerClient: DepotServerClient,
-    sdlcServerClient: SDLCServerClient,
     pluginManager: LegendQueryPluginManager,
     queryId: string,
   ) {
-    super(applicationStore, depotServerClient, sdlcServerClient, pluginManager);
+    super(applicationStore, depotServerClient, pluginManager);
 
     makeObservable<ExistingQueryEditorStore, '_query'>(this, {
       _query: observable,
