+++ packages/legend-application-query/tsconfig.build.json
@@ -11,7 +11,6 @@
     { "path": "../legend-shared/tsconfig.build.json" },
     { "path": "../legend-storage/tsconfig.build.json" },
     { "path": "../legend-server-depot/tsconfig.build.json" },
-    { "path": "../legend-server-sdlc/tsconfig.build.json" },
     { "path": "../legend-application/tsconfig.build.json" },
     { "path": "../legend-art/tsconfig.build.json" }
   ]
