+++ packages/legend-application-query/tsconfig.json
@@ -13,7 +13,6 @@
     { "path": "../legend-shared" },
     { "path": "../legend-storage" },
     { "path": "../legend-server-depot" },
-    { "path": "../legend-server-sdlc" },
     { "path": "../legend-application" },
     { "path": "../legend-art" }
   ]
