+++ packages/legend-application-studio/src/stores/LegendStudioBaseStore.ts
@@ -14,14 +14,31 @@
  * limitations under the License.
  */
 
-import { type GeneratorFn, ActionState } from '@finos/legend-shared';
+import {
+  type GeneratorFn,
+  type PlainObject,
+  HttpStatus,
+  NetworkClientError,
+  ActionState,
+  LogEvent,
+  assertErrorThrown,
+} from '@finos/legend-shared';
 import {
   type ApplicationStore,
+  ActionAlertActionType,
+  ActionAlertType,
   ApplicationTelemetry,
-  LegendApplicationSDLCSetupState,
 } from '@finos/legend-application';
-import { flow, flowResult, makeObservable } from 'mobx';
-import type { SDLCServerClient } from '@finos/legend-server-sdlc';
+import {
+  action,
+  computed,
+  flow,
+  flowResult,
+  makeObservable,
+  observable,
+} from 'mobx';
+import { User, SDLCServerClient } from '@finos/legend-server-sdlc';
+import { LEGEND_STUDIO_APP_EVENT } from './LegendStudioAppEvent.js';
 import type { DepotServerClient } from '@finos/legend-server-depot';
 import type { LegendStudioPluginManager } from '../application/LegendStudioPluginManager.js';
 import type { LegendStudioApplicationConfig } from '../application/LegendStudioApplicationConfig.js';
@@ -37,11 +54,14 @@ export type LegendStudioApplicationStore = ApplicationStore<
 
 export class LegendStudioBaseStore {
   applicationStore: LegendStudioApplicationStore;
+  sdlcServerClient: SDLCServerClient;
   depotServerClient: DepotServerClient;
   pluginManager: LegendStudioPluginManager;
 
   initState = ActionState.create();
-  applicationSDLCSetupState: LegendApplicationSDLCSetupState;
+
+  isSDLCAuthorized = false;
+  SDLCServerTermsOfServicesUrlsToView: string[] = [];
 
   constructor(
     applicationStore: LegendStudioApplicationStore,
@@ -49,23 +69,23 @@ export class LegendStudioBaseStore {
     depotServerClient: DepotServerClient,
     pluginManager: LegendStudioPluginManager,
   ) {
-    makeObservable(this, {
+    makeObservable<LegendStudioBaseStore, 'initializeSDLCServerClient'>(this, {
+      isSDLCAuthorized: observable,
+      SDLCServerTermsOfServicesUrlsToView: observable,
+      needsToAcceptSDLCServerTermsOfServices: computed,
       initialize: flow,
+      initializeSDLCServerClient: flow,
+      dismissSDLCServerTermsOfServicesAlert: action,
     });
 
     this.applicationStore = applicationStore;
-    this.applicationSDLCSetupState = new LegendApplicationSDLCSetupState(
-      applicationStore,
-      sdlcServerClient,
-    );
+    this.sdlcServerClient = sdlcServerClient;
     this.depotServerClient = depotServerClient;
 
     this.pluginManager = pluginManager;
 
     // Register plugins
-    this.applicationSDLCSetupState.sdlcServerClient.setTracerService(
-      this.applicationStore.tracerService,
-    );
+    this.sdlcServerClient.setTracerService(this.applicationStore.tracerService);
     this.depotServerClient.setTracerService(
       this.applicationStore.tracerService,
     );
@@ -81,15 +101,26 @@ export class LegendStudioBaseStore {
     this.initState.inProgress();
 
     // setup SDLC server client
-    yield flowResult(
-      this.applicationSDLCSetupState.initializeSDLCServerClient(),
-    );
+    yield flowResult(this.initializeSDLCServerClient());
+
+    let currentUserID = UNKNOWN_USER_ID;
+    try {
+      const currentUser = User.serialization.fromJson(
+        (yield this.sdlcServerClient.getCurrentUser()) as PlainObject<User>,
+      );
+      this.sdlcServerClient.setCurrentUser(currentUser);
+      currentUserID = currentUser.userId;
+    } catch (error) {
+      assertErrorThrown(error);
+      this.applicationStore.log.error(
+        LogEvent.create(LEGEND_STUDIO_APP_EVENT.SDLC_MANAGER_FAILURE),
+        error,
+      );
+      this.applicationStore.notifyWarning(error.message);
+    }
 
     // setup telemetry service
-    this.applicationStore.telemetryService.setUserId(
-      this.applicationSDLCSetupState.sdlcServerClient.currentUser?.userId ??
-        UNKNOWN_USER_ID,
-    );
+    this.applicationStore.telemetryService.setUserId(currentUserID);
 
     ApplicationTelemetry.logEvent_ApplicationInitialized(
       this.applicationStore.telemetryService,
@@ -108,5 +139,101 @@ export class LegendStudioBaseStore {
     ).notify_ApplicationLoaded();
 
     this.initState.complete();
+  }
+
+  private *initializeSDLCServerClient(): GeneratorFn<void> {
+    try {
+      this.isSDLCAuthorized =
+        (yield this.sdlcServerClient.isAuthorized()) as boolean;
+      if (!this.isSDLCAuthorized) {
+        this.applicationStore.navigator.jumpTo(
+          SDLCServerClient.authorizeCallbackUrl(
+            this.applicationStore.config.sdlcServerUrl,
+            this.applicationStore.navigator.getCurrentLocation(),
+          ),
+        );
+      } else {
+        // Only proceed intialization after passing authorization check
+
+        // check terms of service agreement status
+        this.SDLCServerTermsOfServicesUrlsToView =
+          (yield this.sdlcServerClient.hasAcceptedTermsOfService()) as string[];
+        if (this.SDLCServerTermsOfServicesUrlsToView.length) {
+          this.applicationStore.setActionAlertInfo({
+            message: `Please read and accept the SDLC servers' terms of service`,
+            prompt: `Click 'Done' when you have accepted all the terms`,
+            type: ActionAlertType.CAUTION,
+            actions: [
+              {
+                label: 'See terms of services',
+                default: true,
+                handler: (): void =>
+                  this.SDLCServerTermsOfServicesUrlsToView.forEach((url) =>
+                    this.applicationStore.navigator.openNewWindow(url),
+                  ),
+                type: ActionAlertActionType.PROCEED,
+              },
+              {
+                label: 'Done',
+                type: ActionAlertActionType.PROCEED_WITH_CAUTION,
+                handler: (): void => {
+                  this.dismissSDLCServerTermsOfServicesAlert();
+                  this.applicationStore.navigator.reload();
+                },
+              },
+            ],
+          });
+        }
+
+        // fetch server features config
+        yield this.sdlcServerClient.fetchServerFeaturesConfiguration();
+      }
+    } catch (error) {
+      assertErrorThrown(error);
+      if (
+        // eslint-disable-next-line no-process-env
+        process.env.NODE_ENV === 'development' &&
+        error instanceof NetworkClientError &&
+        error.response.status === HttpStatus.UNAUTHORIZED
+      ) {
+        this.applicationStore.setActionAlertInfo({
+          message:
+            'The first time the application starts in development mode, the developer would need to authenticate using SDLC server. Please do so then manually reload the app',
+          type: ActionAlertType.STANDARD,
+          actions: [
+            {
+              label: 'Authenticate using SDLC',
+              type: ActionAlertActionType.PROCEED,
+              default: true,
+              handler: (): void => {
+                this.applicationStore.navigator.openNewWindow(
+                  this.sdlcServerClient.currentUserUrl,
+                );
+                this.applicationStore.setBlockingAlert({
+                  message:
+                    'Waiting for the developer to authenticate using SDLC server',
+                  prompt:
+                    'Please manually reload the application after authentication',
+                });
+              },
+            },
+          ],
+        });
+      } else {
+        this.applicationStore.log.error(
+          LogEvent.create(LEGEND_STUDIO_APP_EVENT.SDLC_MANAGER_FAILURE),
+          error,
+        );
+        this.applicationStore.notifyError(error);
+      }
+    }
+  }
+
+  get needsToAcceptSDLCServerTermsOfServices(): boolean {
+    return Boolean(this.SDLCServerTermsOfServicesUrlsToView.length);
+  }
+
+  dismissSDLCServerTermsOfServicesAlert(): void {
+    this.SDLCServerTermsOfServicesUrlsToView = [];
   }
 }
