+++ packages/legend-application/package.json
@@ -45,7 +45,6 @@
   "dependencies": {
     "@finos/legend-art": "workspace:*",
     "@finos/legend-graph": "workspace:*",
-    "@finos/legend-server-sdlc": "workspace:*",
     "@finos/legend-shared": "workspace:*",
     "@types/css-font-loading-module": "0.0.7",
     "@types/papaparse": "5.3.3",
