+++ packages/legend-application/tsconfig.build.json
@@ -8,7 +8,6 @@
   "references": [
     { "path": "../legend-shared/tsconfig.build.json" },
     { "path": "../legend-art/tsconfig.build.json" },
-    { "path": "../legend-server-sdlc/tsconfig.build.json" },
     { "path": "../legend-graph/tsconfig.build.json" }
   ]
 }
