+++ packages/legend-extension-dsl-data-space/src/stores/query/DataSpaceQueryEditorStore.ts
@@ -43,7 +43,6 @@ import {
   QUERY_PROFILE_TAG_DATA_SPACE,
 } from '../../DSLDataSpace_Const.js';
 import { getDataSpace } from '../../graphManager/DSLDataSpace_GraphManagerHelper.js';
-import type { SDLCServerClient } from '@finos/legend-server-sdlc';
 
 const createQueryDataSpaceTaggedValue = (
   dataSpacePath: string,
@@ -67,7 +66,6 @@ export class DataSpaceQueryEditorStore extends QueryEditorStore {
   constructor(
     applicationStore: LegendQueryApplicationStore,
     depotServerClient: DepotServerClient,
-    sdlcServerClient: SDLCServerClient,
     pluginManager: LegendQueryPluginManager,
     groupId: string,
     artifactId: string,
@@ -77,7 +75,7 @@ export class DataSpaceQueryEditorStore extends QueryEditorStore {
     runtimePath: string | undefined,
     executionKey: string | undefined,
   ) {
-    super(applicationStore, depotServerClient, sdlcServerClient, pluginManager);
+    super(applicationStore, depotServerClient, pluginManager);
 
     this.groupId = groupId;
     this.artifactId = artifactId;
