+++ packages/legend-extension-dsl-data-space/tsconfig.build.json
@@ -12,7 +12,6 @@
     { "path": "../legend-application-studio/tsconfig.build.json" },
     { "path": "../legend-application-query/tsconfig.build.json" },
     { "path": "../legend-server-depot/tsconfig.build.json" },
-    { "path": "../legend-server-sdlc/tsconfig.build.json" },
     { "path": "../legend-shared/tsconfig.build.json" },
     { "path": "../legend-storage/tsconfig.build.json" },
     { "path": "../legend-application/tsconfig.build.json" },
